should not be able to remove himself

+++ controllers/user.js
@@ -51,7 +51,9 @@ async function deleteUser(req, res) {
     if (!userId) {
       throw {status: 422, message: 'Id is required'}
     }
-
+    if (req && req.user && req.user.external_id == userId) {
+      throw {status: 422, message: 'You cannot delete yourself'}
+    }
     const deleteCount = await db.user.destroy({where: {external_id: userId}})
     res.status(deleteCount > 0 ? 204 : 404).send();
   } catch (err) {
